Script to run all examples in parallel. [skip ci]
