--- .travis/deploy_gh_pages.sh.in
@@ -1,63 +1,0 @@
-#!/usr/bin/env bash
-README=@README_TO_CHANGE@
-FAILEDTXT=$(cat FAILED)
-SUCCESSTXT=$(cat SUCCEEDED)
-
-SUCCESSTXT=""
-if [[ -f SUCCEEDED ]]; then
-    SUCCESSTXT=$(cat SUCCEEDED)
-fi
-
-BROKENTXT=""
-if [[ -f BROKEN ]]; then
-    BROKENTXT=$(cat BROKEN)
-fi
-
-BLACKLISTEDTXT=''
-if [[ -f BLACKLISTED ]]; then
-    #statements
-    BLACKLISTEDTXT=$(cat BLACKLISTED)
-fi
-
-cat <<-EOF > $README 
-[![Build Status](https://travis-ci.org/BhallaLab/moose-examples.svg?branch=master)](https://travis-ci.org/BhallaLab/moose-examples).
-
-Examples, tutorial and demo scripts.
-
-# Status of scripts tested by Travis.
-
-## Successful scripts
-
-These scripts ran successfully.
-
-$SUCCESSTXT
-
-## Failed scripts
-
-These scripts failed to execute successfully on Travis.
-$FAILEDTXT
-
-## Broken scripts
-
-These scripts are marked "BROKEN" by one or more developers.
-
-$BROKENTXT
-
-## Blacklisted scripts
-
-These scripts were not tested because they require human interaction :-(.
-
-$BLACKLISTEDTXT
-
-EOF
-
-# If there is any text in FAILEDTXT then exit 1
-FAILEDTXT_WITHOUT_WHITESPACE=$(echo "$FAILEDTXT" | xargs)
-if [ -z "$FAILEDTXT_WITHOUT_WHITESPACE" ]; then
-    echo "Success."
-    exit 0
-else
-    echo "One of more scripts are breaking. See the output below:"
-    echo "$FAILEDTXT"
-    exit 1
-fi
